requirements.txt
bittensor-cli>=8.2.0rc10
starlette>=0.30.0
pydantic>=2
torch>=2
numpy>=1
setuptools>=68
tenacity>=9
openai>=1.47
pytest-flaky>=3.8.1
